Modularize package fetching
Introduce a parent class to keep the interface consistent between sources. To
start with, that is just the Condaforge channel and Pypi. Next steps will be
testing this and accounting for edge-cases (e.g. "py-" prefixes), followed by
seeing what other sources can be checked for releases, like Github or Bitbucket.
These would likely depend on the existing recipe, if the download or homepage
URLs point to a valid project page. However, this would also add overhead in
that it would require fetching the recipe file itself.