Implement basic functionality into main.py
Refactor the Github code into the main class for the moment. Not sure if this
really needs to be separate or not, since conda-forge is on Github. If that
changes then this needs updating anyways.

Currently supports listing packages by name as arguments, as well as by Github.
For Github, it uses a personal (read-only) token, and uses all conda-forge
repositories that you have push access to as input (stripping "-feedstock" from
their names).