Refactor test mocking
Complete redo of how mocking is handled, now using OOP and context managers.
This is a proper wrapper of RequestsMock that gives much more flexibility and
allows multiple mocks to take place at once.

Also remove the general dependency of mock_repodata which was too anaconda
specific. There is still a hard-coded nature of the mock package list, but
the format is more generic.