Optimize GitHub further
For some reason `repo.owner.name` is way slower than `repo.owner.login`. I'd
imagine the former is actually making another HTTP request.

GitHub responded and recommended the use of the `/user/repos` endpoint to do
what we want, so I am honoring that here, and with the other change this is
still faster than previously, and for users that don't have an insane amount of
repositories, this should actually be quite quick.